+++ protocols/mdns/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.47.1
+
+- Fixed a Arithmetic overflow bug in the mDNS implementation method processing response packets with extremely large TTL values.
+  See [PR 5967](https://github.com/libp2p/rust-libp2p/pull/5967)
+
 ## 0.48.0
 
 - Remove `async_std` dependency [PR 5958](https://github.com/libp2p/rust-libp2p/pull/5958)